Fix bug that breaks azurerm backend on tf v1.1.x
To maintain compatibility with older versions of terraform, the azurerm
backend code checks the version of terraform in use, and modifies the
names of some backend config auth parameters. Unfortunately this check
is faulty, and it only works with current 1.x versions because hashicorp
haven't released a build with major _and_ minor numbers greater than 0.

This code updates this check to actually test the major build number,
and where testing equality it relies on type checks and strict equality
checking to ensure that we aren't treating numeric values as booleans.