ensure backend fails if keys are not allowed (comments)
I think this feature is not worth the pain of trying to keep it going.  If users of this extension need to make sure the storage account exists they can simply put this in their pipeline.  Then they can create exactly what they need, exactly how they need it with the right security settings.